move dichotomous model averaging structs
