+++ bmds/bmds3/types/structs.py
@@ -1,10 +1,14 @@
 import ctypes
 from textwrap import dedent
-from typing import NamedTuple
+from typing import List, NamedTuple
 
 import numpy as np
 
 from .. import constants
+from .common import list_t_c
+
+# DICHOTOMOUS MODELS
+# ------------------
 
 
 class DichotomousAnalysisStruct(ctypes.Structure):
@@ -169,4 +173,68 @@ class DichotomousStructs(NamedTuple):
             Summary:
             {self.summary}
             """
+        )
+
+
+# DICHOTOMOUS MODEL AVERAGING
+# ---------------------------
+
+
+class DichotomousMAAnalysisStruct(ctypes.Structure):
+    _fields_ = [
+        ("nmodels", ctypes.c_int),
+        ("priors", ctypes.POINTER(ctypes.POINTER(ctypes.c_double)),),
+        ("nparms", ctypes.POINTER(ctypes.c_int)),
+        ("actual_parms", ctypes.POINTER(ctypes.c_int)),
+        ("prior_cols", ctypes.POINTER(ctypes.c_int),),
+        ("models", ctypes.POINTER(ctypes.c_int)),
+        ("modelPriors", ctypes.POINTER(ctypes.c_double)),
+    ]
+
+    @classmethod
+    def from_python(cls, models: List[DichotomousAnalysisStruct]):
+
+        # list of floats
+        priors = [
+            list_t_c(model.prior[: model.parms * model.prior_cols], ctypes.c_double,)
+            for model in models
+        ]
+
+        # pointer of pointers
+        priors2 = list_t_c(
+            [ctypes.cast(el, ctypes.POINTER(ctypes.c_double)) for el in priors],
+            ctypes.POINTER(ctypes.c_double),
+        )
+
+        return cls(
+            nmodels=ctypes.c_int(len(models)),
+            priors=priors2,
+            nparms=list_t_c([model.parms for model in models], ctypes.c_int),
+            actual_parms=list_t_c([model.parms for model in models], ctypes.c_int),
+            prior_cols=list_t_c([model.prior_cols for model in models], ctypes.c_int),
+            models=list_t_c([model.model for model in models], ctypes.c_int),
+            modelPriors=list_t_c([1 / len(models)] * len(models), ctypes.c_double),
+        )
+
+
+class DichotomousMAResultStruct(ctypes.Structure):
+    _fields_ = [
+        ("nmodels", ctypes.c_int),
+        ("models", ctypes.POINTER(ctypes.POINTER(DichotomousModelResultStruct))),
+        ("dist_numE", ctypes.c_int),
+        ("post_probs", ctypes.POINTER(ctypes.c_double)),
+        ("bmd_dist", ctypes.POINTER(ctypes.c_double)),
+    ]
+
+    @classmethod
+    def from_python(cls, models: List[DichotomousModelResultStruct]):
+        _results = [ctypes.pointer(model) for model in models]
+        nmodels = len(models)
+        dist_numE = 200
+        return DichotomousMAResultStruct(
+            nmodels=nmodels,
+            models=list_t_c(_results, ctypes.POINTER(DichotomousModelResultStruct)),
+            dist_numE=ctypes.c_int(dist_numE),
+            post_probs=(ctypes.c_double * nmodels)(),
+            bmd_dist=(ctypes.c_double * (dist_numE * 2))(),
         )
